feat: field -> category로 명칭 변경
