+++ components/main/lectureCategory/category/style.ts
