+++ components/main/lectureCategory/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Field from "./field";
-import { FieldWrapper, StyledRoot } from "./style";
+import Category from "./category";
+import { CategoryWrapper, StyledRoot } from "./style";
 
 const dummy = [
   { id: 1, name: "개발" },
@@ -11,18 +11,18 @@ const dummy = [
   { id: 6, name: "기타" },
 ];
 
-const MainLectureFields = () => {
+const MainLectureCategory = () => {
   return (
     <StyledRoot>
       <h3>내가 찾고 싶은 강의 분야는?</h3>
       <small>10초면 내게 맞는 강의를 찾을 수 있어요.</small>
-      <FieldWrapper>
-        {dummy.map((field) => (
-          <Field key={field.id}>{field.name}</Field>
+      <CategoryWrapper>
+        {dummy.map((category) => (
+          <Category key={category.id}>{category.name}</Category>
         ))}
-      </FieldWrapper>
+      </CategoryWrapper>
     </StyledRoot>
   );
 };
 
-export default MainLectureFields;
+export default MainLectureCategory;
